requirements.txt
beautifulsoup4>=4.12.3
openai>=1.51.0
openpyxl>=3.1.5
pandas>=2.2.3
python-dotenv>=1.0.1
requests>=2.32.3
psycopg2-binary>=2.9.10
SQLAlchemy>=2.0.36
streamlit>=1.39.0
